refactor(header): drop const wrapper files for logo contract
Colocate crest SVG and HEADER_* exports on LogoHomeLink; Header script imports the same module. Removes header-dom.ts and logo-crest.ts.
