+++ src/components/LogoHomeLink.tsx
@@ -1,8 +1,19 @@
 import { useId, useSyncExternalStore, type ReactNode } from 'react'
 import { motion, useReducedMotion, type Transition, type Variants } from 'motion/react'
-import { HEADER_MOBILE_MENU_OPEN_ATTR, HEADER_SELECTOR } from '@/consts/header-dom'
-import { LOGO_CREST, LOGO_HOME_ARIA } from '@/consts/logo-crest'
 import { $ } from '@/lib/dom-selector'
+
+/** Shared with `Header.astro` script — menu open toggles this on `[data-header]`. */
+export const HEADER_SELECTOR = '[data-header]'
+export const HEADER_MOBILE_MENU_OPEN_ATTR = 'data-mobile-menu-open'
+
+/** Crest SVG for this link only; hero uses `/public/logo.webp`. */
+export const LOGO_CREST = {
+  viewBox: '0 0 303 303',
+  path: 'm188.1 150.2l90.6 137.4-34.9-16c-7.7-3.5-15.1-7.7-21.9-12.3-15.3-10.4-29-25.3-39.6-43.1l-17.8-30.2-12 18.2-12-18.2-17.7 30.2c-10.6 17.8-24.4 32.7-39.7 43.1-6.8 4.6-14.2 8.8-21.9 12.3l-34.9 16 90.6-137.4-90.6-137.4 34.9 16c7.7 3.5 15.1 7.7 21.9 12.3 15.3 10.4 29.1 25.3 39.7 43.1l17.7 30.2 12-18.2 12 18.2 17.8-30.2c10.6-17.8 24.3-32.7 39.6-43.1 6.8-4.6 14.2-8.8 21.9-12.3l34.9-16zm-21-31.8l18.4 27.9 81.3-123.3-21.2 9.7c-7.5 3.5-14.7 7.5-21.3 12-14.8 10-28 24.5-38.3 41.8zm-128.9-95.4l81.2 123.2 18.4-27.8-18.8-32c-10.3-17.3-23.6-31.7-38.4-41.8-6.6-4.5-13.8-8.5-21.3-11.9zm147.3 131.1l-23.5-35.6-9.5-14.5-9.6 14.5-23.4 35.6-81.3 123.2 21.2-9.7c7.5-3.4 14.6-7.4 21.2-11.9 14.8-10.1 28.1-24.5 38.4-41.8l19.5-33.1 14-23.7 13.9 23.7 19.5 33.1c10.3 17.3 23.6 31.7 38.4 41.8 6.6 4.5 13.8 8.5 21.3 11.9l21.1 9.7zm-23.4 27.8l-9.6-16.2-9.5 16.2 9.5 14.4z',
+  restFill: 'oklch(0.765 0.038 55.5)',
+} as const
+
+const LOGO_HOME_ARIA = 'La Velada del Año VI - Inicio'
 
 const EASE_OUT = [0.23, 1, 0.32, 1] as const
 const EASE_FILM = [0.77, 0, 0.175, 1] as const
