--- src/consts/header-dom.ts
@@ -1,3 +1,0 @@
-/** DOM contract between `sections/Header.astro` and `LogoHomeLink`. */
-export const HEADER_SELECTOR = '[data-header]'
-export const HEADER_MOBILE_MENU_OPEN_ATTR = 'data-mobile-menu-open'
