+++ src/sections/Header.astro
@@ -181,7 +181,7 @@ const scrolledBar =
 </header>
 
 <script>
-  import { HEADER_MOBILE_MENU_OPEN_ATTR, HEADER_SELECTOR } from '@/consts/header-dom'
+  import { HEADER_MOBILE_MENU_OPEN_ATTR, HEADER_SELECTOR } from '@/components/LogoHomeLink.tsx'
   import { $, $$ } from '@/lib/dom-selector'
 
   let headerAbort: AbortController | null = null
